+++ packages/discord.js/src/structures/MediaChannel.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const BaseForumChannel = require('./BaseForumChannel');
+
+/**
+ * Represents a media channel.
+ * @extends {BaseForumChannel}
+ */
+class MediaChannel extends BaseForumChannel {}
+
+module.exports = MediaChannel;
